Creates CI for some PR validation
If a PR is opened and targets the `main` branch, this creates a job to run `npm --clean-install` on the code in the PR, thus running all the validation that the openrpc tooling provides. If the job exits with a non-zero code, the job fails and the merge is not allowed.